Integrated Conduit into the Voice SDK, Separated Conduit into its own package, Restructured code, Code clean up, Integrated Conduit into the Voice SDK, Separated Conduit into its own package, Restructured code, Code clean up
Summary:
This is the initial integration of Conduit into the Voice SDK.

Factored out Conduit code out of the Wit package.

Restructured code to put the Conduit project as a separate library.

Cleaned up code to remove Assistant-specific names.

This is the initial integration of Conduit into the Voice SDK.

Factored out Conduit code out of the Wit package.

Restructured code to put the Conduit project as a separate library.

Cleaned up code to remove Assistant-specific names.

Reviewed By: yolanother

Differential Revision: D36257513

fbshipit-source-id: fd1a970158fb4d88584fe3af892dee41e03ee5bd