Add VoiceServiceReference and dedicated GazeVoiceActivator
Summary: Adds a gaze activator that will automatically toggle voice service activation. This is adding a VoiceServiceReference that will either find the current voice service in the scene or use a selected voice service. It provides an easier way for devs to point a voice component at a voice service.

Reviewed By: MokhtarKhorshid, ryanbailey-fb

Differential Revision: D36615449

fbshipit-source-id: 50aa15e81e6d9b4021e5267bda44e543c8448b16