Improve messaging when no app id is available to server token
Summary: When the /apps request doesn't return any apps associated with a server token we were failing silently. Furthermore if an app was already loaded in the WitConfig it would leave all of that app's data even though it didn't match the server/client token. This clears that data and shows an error if your application is in this state.

Reviewed By: patapizza

Differential Revision: D38134513

fbshipit-source-id: 1e922b9deddb22f894de6b5e613dfa4439b2950c