Conduit Entity Synchronizer
Summary:
Added the entity synchronizer code that synchronizes Wit.Ai entities with local Enum in code.
This code adds the bulk of the code, but without the UI component.
Due to logistical challenges, this diff will require some extra revisions before being ready to commit. I am committing this early to unblock other devs and allow them to commandeer the diff to landing.

Reviewed By: MokhtarKhorshid

Differential Revision: D38558739

fbshipit-source-id: f99acfb74561bd0e7188588e4d15f7d32d2c0e0c