Move updates checks to main thread
Summary: When using audio files for input audio is fed into the audio source on a secondary thread. Time.time throws an exception if it is read off the main thread. We'll use the time values from the last update to determine last word time instead to protect audio data thread.

Reviewed By: maidaneze

Differential Revision: D40499422

fbshipit-source-id: 440f2870180e5efa02f965b20559563dc3b76e60