Fix issues with app focus and mic
Summary:
A prototyper was having issues with the mic stopping recording if they used the editor utilities and had alwaysRecord turned on in AudioBuffer. It turns out we have an optimization to stop the mic if the application loses focus. This is triggered even if you are clicking UI in the inspector within unity.

Fix:
We're only going to check application focus on Android devices and never while in the editor. This will allow Quest devices to still release the mic if the user removes the headset.

Reviewed By: ryanbailey-fb

Differential Revision: D40655551

fbshipit-source-id: dec1622371e60de521045ce15c83fc2c19b55e23