Fix unhandled exception in Abort
Summary: If a dev called DeactivateAndAbort() sometimes the Abort would throw an exception deep in the HttpWebRequest. This was happening because we were calling web methods after abort was called. This change sets the status to aborted before calling abort and checks the status before making calls to HttpWebRequest.

Reviewed By: MokhtarKhorshid

Differential Revision:
D42005132

LaMa Project: L1079626

fbshipit-source-id: f5f237e968af83f5955670701285ef0343c9b39d