Add AudioDurationTracker in wit-unity
Summary: This change adds a class called `AudioDurationTracker` in wit-unity. The ADT will keep track of the count of audio bytes that are being sent to Wit.ai. At the end of the speech, the ADT calculates the total length of audio and logs that information. This information will be useful to calculate the end-to-end latency of Voice SDK requests.

Reviewed By: maidaneze, ryanbailey-fb

Differential Revision:
D41981081

Privacy Context Container: L1115215

fbshipit-source-id: bad8268a5f41a556c2e3472ff922bc513b225025