Add Composer's Context_map params to WitConfig
Summary:
Design doc [here](https://fburl.com/gdoc/khh5im05)
* added Export endpoint call
* added download of zip file
* added new call in `WitAppInfoUtility.cs` for running the above
* parsing of zip file
* adding of Composer contents to appInfo

+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬+¬++¬

It would be good to split this *entirely* out into the composer plugin, but that'll require a 3rd diff, but that is not easily possible.
The best approach would be to make `WitAppInfo` a _partial_ and then define the Composer part in the Composer package.... however, those are different assemblies and we cannot split partials across assemblies. :(

That said, presently the composer data in the export is already going to be in the export download from the server whether they've got the composer plugin or not.  To make the class itself dynamic, it wouldn't be able serializable.
.... so some of that stuff has to remain in the main package. :(

Reviewed By: ryanbailey-fb

Differential Revision: D43737688

fbshipit-source-id: 5c7e52c5164cdfa0dfb417cda439e6595e5240a9