Deactivating request after audio clip has been sent
Summary:
The voice commands for the GauntletTestApp weren't closing the request. This meant the server would keep waiting for the max request duration without closing the request.
Calling the Appropiate VoiceExperience Deactivate function in order to close requests as soon as the audio clip is processed.

Reviewed By: ryanbailey-fb

Differential Revision:
D43979920

Privacy Context Container: L1079626

fbshipit-source-id: 7b253c203b621069477b6187ea594eb41533f6a2