Adding csc.rsp file to wit
Summary:
Apparently, putting an `rsp` next to an `asmdef` will have it also imported, [according to this post](https://forum.unity.com/threads/rsp-file-per-assembly-or-folder.572758/).

> Placing a RSP file next to a Asmdef will then be picked up for that specific Asmdef compilation.
If you have a RSP file in the Asset folder, it will not be used for Asmdef's having RSP files next to them.

Reviewed By: ryanbailey-fb

Differential Revision:
D43981607

Privacy Context Container: L1079626

fbshipit-source-id: 9c7cc2b6fa750b99e51404b8e0ed80ce124e4376