Turn off looping in TTS
Summary:
If the audio source used for the speaker is being used for other audio as well (such as a phone receiver which needs to play a dialtone), it is possible that the dev sets the audio source to loop.   We weren't checking for that before playing, meaning that we'd just always loop the audio.

Also ensured that we skip trying to speak if there is no text to speak.

Reviewed By: yolanother

Differential Revision: D46800542

fbshipit-source-id: 83c73b16347a3a46fca19097cb97524140b548d5