Add stub methods for speech block control
Summary: Adds api to the ISpeaker interface to handle notifying the TTS service about sending related data. For example, you'd send StartTextBlock(), then individual requests for several sentences then EndTextBlock(). Also adds a method to notify the TTS system that speech is about to start. This can be used to prime a connection to the speech system when needed and is fired at the beginning of an activation.

Reviewed By: ryanbailey-fb

Differential Revision:
D48524050

Privacy Context Container: L1079626

fbshipit-source-id: 2d7dd7d322009cd57de7cc3878fff4cd53c23d7a