Fix leaky AudioBuffer Singleton
Summary:
The Singleton in `Meta.WitAi.Data.AudioBuffer` leaks GameObjects if it is ever called from `OnDisable` after its `OnDestroy` is called. Since lifecycle events are only guaranteed to be in the correct order within a class, this currently happens very commonly. Specifically:

1. The `AudioBuffer` instance is destroyed
2. `WitService`'s `OnDisable()` is called, which in turn calls `SetMicDelegates`, which in turn gets `AudioBuffer.Instance`
3. When `get`ting `AudioBuffer.Instance`, `Application.isPlaying` is checked, which is `true` even during "shutdown"

This diff fixes the issue by caching `_isApplicationQuitting`, updating it during `OnApplicationQuit` (which is guaranteed to be called beforehand), and checking it during Instance creation.

Reviewed By: Barbog

Differential Revision: D49155673

fbshipit-source-id: c50d768446fc3b43ff97b5fb787ccd81dbaf0784