Add refresh and default view to DropDownAttribute
Summary:
Adds an optional refresh button and defaults to the regular field if no options are found. This allows devs to fill in the field manually if using a field that may have failed to fetch. Using refresh also has a secondary method that it can use to trigger a longer async fetch if needed vs the GUI refresh.

 {F1117411501}

Reviewed By: ryanbailey-fb

Differential Revision: D50209268

fbshipit-source-id: adfda1dc0fcddfa17c7a3d93fdc7ae949a26e7fc