Save Exports
Summary:
Because the server doesn't have an auto-backup function, this implements an auto-save which retains a copy of each new app export we see.

These are saved locally to disk.

Note that we've also removed the IExportParserPlugin interface from the Character/Composer parsers as these are now pull-processors (they request the export themselves) rather than push processors which are handed the zip archive whenever it is received.
(both work, but we may as well only do one)

Reviewed By: ryanbailey-fb

Differential Revision: D50790836

fbshipit-source-id: cbbae5715046b587e3fd11591bac6944cba525eb