Perception and Direction systems
Summary:
Adds two new systems for handling different types of dynamic context.

A) the perception system
B) the direction system

 {F1143700134}
In this diff, they only display the perceptions and directions when in Play mode.

You can also add and remove context whilst in play mode, to help test it and understand the system.

Reviewed By: yolanother

Differential Revision: D50984063

fbshipit-source-id: 28ad43ddc57873963d52878036fc8867fdb6e856