Make Reserved Path classes
Summary:
Because NPCs and Composer are different packages, we need a system which can extend Composer -- notably the context map.
The main thing to extend in this case is the concept of "reserved paths" in the context map. Up until this point, these were just hard-coded string checks.

So, we've added a series of "Reserved Context Map Path" classes.
This includes the new context fields discussed ad nausium in meetings.
eg:
* AddDynamicContext is moved to CurrentContextMap.ImmediateContext

This also includes the changes to ensure that we're not just overwriting the context map whenever we want -- we instead just clear the non-reserved values when we want to 'reset' something.

Reviewed By: yolanother

Differential Revision: D51073297

fbshipit-source-id: 894b3b8946f3f4808a7f7539e9c4d444181d7c31