AI Task Context Node Parser
Summary:
Adds an extension parser for Composer so as to recognize and retain the information about Context Nodes that are actually AI tasks.

It just saves the value as
"AI_Task: <task_name>" in the server variables, and this is later used by the AI Task parser to be replaced with actual values.

Reviewed By: ryanbailey-fb

Differential Revision: D52786840

fbshipit-source-id: 00c6012066cc30aea54998af8767998afc84b007