Telemetry Facade fix
Summary:
The editor for core VoiceSDK asmdef has
```
        {
            "name": "com.meta.xr.sdk.voice.telemetry",
            "expression": "",
            "define": "VSDK_TELEMETRY_AVAILABLE"
        },
```
But the `#if !VSDK_TELEMETRY_AVAILABLE && UNITY_2021_2_OR_NEWER`
inside TelemetryFacade isn't triggering.   Specifically, it looks like the `UNITY_2021_2_OR_NEWER` define isn't necessarily present. This means the `TelemetryFacade` never gets 'loaded' and the lack of the Telemetry package causes errors.

Given we only support 2021+ now, we can remove this tag.
(so, basically, this is just a reversion of D46501185)

Reviewed By: ryanbailey-fb

Differential Revision: D53925237

fbshipit-source-id: 0ae77f7ce5de78a8b614a8d8ebe4ac5b58be6260