Refactored logging
Summary:
This is the first step in refactoring logging to increase utility and performance.

This step maintains backward compatibility and channels logs through VLog as usual, except that it now goes through the new logger. However, the foundational elements of the new logging system as added here. Just not hooked up yet.

Reviewed By: ryanbailey-fb

Differential Revision: D54780281

fbshipit-source-id: 30a2469d3da2407e379b914d7978c8c57a45acce