Hide repeated correlation IDs for the same log messages
Summary:
When a message is repeated, we will only log the first correlation ID the first time, after that it will be replaced with ... to allow it to be collapsed.

If it is repeated with the same ID, then we still show the ID as it will still be collapsible.

Reviewed By: ryanbailey-fb

Differential Revision: D55722667

fbshipit-source-id: 5ecd736678ab561dce9bd54f518bfd475d1426ed