Partial migration of VLog
Summary:
Deprecated a number of the VLog methods in favor of VLogger.
This only removes a small percentage of the log statements we have. Additional follow up work is needed to migrate everything, but this servers as an example.
We still also need to replace string interpolations with structured logging as well, which this change only does in a few examples.

Reviewed By: yolanother

Differential Revision: D56916676

fbshipit-source-id: 8876d950844c3a9b9f53a043bc7ba16198df1142