Fix issue with HasDynamicEntityRegistry
Summary: There was a problem with the singleton implemention of HasDynamicEntityRegistry. It was doing a scan for the registry when this property was called. This component is generally only used after it has been instantiated and enabled. The instance is assigned on enablement so there's generally no need to search for the component if it hasn't been already assigned. We can assume at this phase of a voice activation that it should already be ready or is not in use. If there is need to search, the Instance field will manage it.

Reviewed By: MokhtarKhorshid

Differential Revision:
D57593710

Privacy Context Container: L1079626

fbshipit-source-id: 28e869ad9fd05ae96675340f590d30d0c2829f20