Thread utility should rethrow exceptions
Summary: Not sure what I was thinking on the exceptions here. All we really want to do is capture any unhandled exceptions and pass them to the logger  for tracking. Changing the undefined return behaviors to rethrow the exceptions. I'm worried these were catching ThreadAbortExceptions and preventing the threads from executing.

Reviewed By: ryanbailey-fb

Differential Revision: D57897501

fbshipit-source-id: 7986fcd85948f61dd7175033fc72f6d433047a0a