Synchronize log methods
Summary: Currently methods under the logger are not thread safe and need to have some forced concurrency to work properly. This especially covers things like id generation and the dictionary ring buffer. This is a intirum fix that forces the critical methods to be concurrent. This will come at a potential cost of latency and perf, but will work for now since in prod most of our logging is suppressed by default.

Reviewed By: ryanbailey-fb

Differential Revision: D58023580

fbshipit-source-id: 68aeb30912c35911aff7e482ba9b6dad2b183114