Make tts prep async
Summary: As we make more methods async/await there will be a spiderweb of asyncs that need to happen. In this case there was a get component that was occasionally happening on the background thread of an async. Switched that get component over to an awaited async, but that spiderwebbed out to several other calls. This should be the minimum number of methods that need to be touched  to do so.

Reviewed By: Barbog

Differential Revision:
D58389824

Privacy Context Container: L1079626

fbshipit-source-id: 14204917c5ae55c912fafb4bb1e0b41e0373b43c