Move version tag setup to WitConfiguration
Summary:
Wit version tag is now a full versioning system for the current state of a wit application.  The previous setup process assumed the versioning was for composer only.  Now that this is a full configuration setting we need to move the editor and build version tags up to the WitConfiguration.  More urgently, the web socket implementation requires version tag immediately on authentication which means the version needs to be accessible outside of composer.

Below are the changes made in this diff to address the issues:
1. Adding editorVersionTag and buildVersionTag to WitConfiguration.
2. Adding editor and build version tag dropdowns to WitConfiguration inspector.
3. Deprecating previous request option and composer service based version tagging.
4. Adding tag to web socket authentication.

{F1770543184}

Reviewed By: Barbog

Differential Revision: D60176394

fbshipit-source-id: 6fadf18c79f9fe35396f45d1dc0f65782551b6cb