Add http based two phase with injectable phase transition control
Summary: Adds a prefab for two phase activation that will work with http requests. This also adds in a component based evaluator that will determine if the second phase should be entered immediately after transcription or if the client wants to take over control of sharing (typically via an RPC mechanism).

Reviewed By: ryanbailey-fb

Differential Revision: D60920356

fbshipit-source-id: af66b53d6244f8224b6ae2c46e53d8af6d078da1