_log is not defined anywhere in the file - breaks webgl builds
Summary:
One of the libraries we rely on, Wit, has a minor bug that stops web assembly builds from succeeding. This diff fixes that.

`_log` variable seems to be a left-over from a refactor. There is no variable of that name in the class, and no global of that name either. Changing it to `Logger.Warning` following other patterns in the code fixes the issue.

Reviewed By: jermenkoo

Differential Revision: D65243496

fbshipit-source-id: d14137283e350951ef473976d139ba2429be5a5d