Add error handler for async void operations
Summary: Adds a new extension to our TaskExtensions to handle void async calls and log any exceptions that may have happened in the event that async got move dto another thread. We will transition to using this pattern instead of the _ = Async() pattern from now on so we don't lose exceptions.

Differential Revision: D69885556

fbshipit-source-id: 29072289e0fc2aacd007b60127d5545ef77b6447