Better handle async timeouts so we catch exceptions
Summary: Creates a new TaskExtensions extension Task.TimeoutAfter(ms). This extension handles timing out a task as we did in specific code but creates a reusable utility. It also emits exceptions if the task being run against the timeout threw an exception for some reason. We were previously swallowing the exception as the code was written before.

Reviewed By: ryanbailey-fb

Differential Revision:
D70140402

Privacy Context Container: L1079626

fbshipit-source-id: 895821b2b8669da5e9df037eb9e22e7ac9f40874