Cancel ongoing dictation when app stops
Summary:
Fix for T225986254 - cancels ongoing dictation in AppDictationExperience.OnDisable, so that the service will be properly shut down.

Also, adds a check in Wit.ai AudioBuffer.StopUpdateVolume to avoid calling StopCoroutine if the object is already being destroyed, since that was causing an exception which interrupted the cancellation process.

Reviewed By: ryanbailey-fb, yolanother

Differential Revision: D76282929

fbshipit-source-id: 2e5fc791714e2c3201b2f95a984ba3c2dbd210f1