Add tooltip box to TTSWit to clarify ambiguous UI
Summary: Fix for T227967671 - the dropdown to add presets, and the fact that it's visually more prominent than the (usually collapsed) list above it, makes it appear that you're setting the current voice by changing it.  Added a tooltip box to clarify the dropdown's purpose and indicate where the setting can be changed (TTSSpeaker).

Reviewed By: yolanother

Differential Revision: D76855833

fbshipit-source-id: af8c9e108ec2803b3c10bd1c0c138b2f6f0f07b2