Ensure auth settings are applied
Summary: Adds a call to refresh auth parameters once we have fully initialized the wit npc platform adapter. This ensures that all of the changes we made are fully applied on startup fixing an issue where the npc was not receiving the auth parameters. Also moves the auth/debug setup before fully config so we don't refresh auth multiple times.

Differential Revision: D77463235

fbshipit-source-id: a66b6c0e37b1fb2bb41d13047068d9ab2c08cf5d