Potential fix for T228086962
Summary: Based on the logs of the error's occurrence (see the task for link to Logview), the issue appears to stem from MuteBlocker trying to create an instance of AudioBuffer as part of disabling itself.  Since that's only used to unsubscribe (which it won't be if this instance is newly created), this changes it to check and not create an instance in that situation.

Reviewed By: yolanother

Differential Revision: D77876091

fbshipit-source-id: 46fc515781da7364073a4b0fb66c10ed75a7b7a3