Potential fix for T228903369
Summary:
Based on testing, the lost words issue appears to originate from large periods of silence being sent to Wit, which it tries to parse (usually as "and") and which interfere with the actual words being parsed after a long enough pause.

This adds a "silence skipping" feature to AudioBuffer, which doesn't record silence after a certain amount of time (currently one second), but starts recording again as soon as non-silent input occurs.

Note: This should likely be a setting (to be used with walkie-talkie style mics) rather than the default behavior, hence why this is marked as draft.

Reviewed By: yolanother

Differential Revision: D77901863

fbshipit-source-id: c18b323fa89c32a5a55ee6f96653279163eb2321