Fix TTS when there is no viseme data but we requested it
Summary:
We request event data from the server with a setting that may or not actually end up having events. We should log cases where this happens for telemetry purposes, but it should be treated as a warning instead of an error so TTS streams accordingly. If we return an error at this point it causes TTS events to fire and end immediately even if we have valid audio data.

TODO: There are other cases/errors that we could get from the server where this could happen. We should investigate this and make sure we either stop the audio from playing or delay the error reporting until the playback completes. (insert task here)

Differential Revision: D79588758

fbshipit-source-id: 67ddb33b38e4707c27c63ad4879e71a3bceeada0